fix(std/node/util): move `validateIntegerRange()` into internal utils
Closes denoland/deno#5397
