+++ std/node/util.ts
@@ -68,23 +68,6 @@ export function isPrimitive(value: unknown): boolean {
   );
 }
 
-export function validateIntegerRange(
-  value: number,
-  name: string,
-  min = -2147483648,
-  max = 2147483647,
-): void {
-  // The defaults for min and max correspond to the limits of 32-bit integers.
-  if (!Number.isInteger(value)) {
-    throw new Error(`${name} must be 'an integer' but was ${value}`);
-  }
-  if (value < min || value > max) {
-    throw new Error(
-      `${name} must be >= ${min} && <= ${max}.  Value was ${value}`,
-    );
-  }
-}
-
 import { _TextDecoder, _TextEncoder } from "./_utils.ts";
 
 /** The global TextDecoder */
